+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## HEAD
+
+### Feature
+
+-   Add `learner.in_bag_example_indices(num_examples, tree_idx)` to Random 
+    Forest Learner to retrieve the indices of the in-bag samples used for
+    training each tree.
+
 ## 0.13.0 - 2025-07-15
 
 ### API Changes